make oauth2 optional
