+++ src/main/java/com/nephest/battlenet/sc2/config/security/DelegatingAuthorizationRequestResolver.java
@@ -1,4 +1,4 @@
-// Copyright (C) 2020-2025 Oleksandr Masniuk
+// Copyright (C) 2020-2026 Oleksandr Masniuk
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 package com.nephest.battlenet.sc2.config.security;
@@ -16,6 +16,7 @@ import org.springframework.security.oauth2.core.endpoint.OAuth2ParameterNames;
 import org.springframework.stereotype.Component;
 
 @Component
+@OAuth2
 public class DelegatingAuthorizationRequestResolver
 implements OAuth2AuthorizationRequestResolver
 {
