+++ src/main/java/com/nephest/battlenet/sc2/config/security/OAuth2.java
@@ -1,0 +1,17 @@
+// Copyright (C) 2020-2026 Oleksandr Masniuk
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+package com.nephest.battlenet.sc2.config.security;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.springframework.boot.autoconfigure.security.oauth2.client.ClientsConfiguredCondition;
+import org.springframework.context.annotation.Conditional;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+@Conditional(ClientsConfiguredCondition.class)
+public @interface OAuth2
+{}
